+++ addon/services/hifi.js
@@ -644,7 +644,7 @@ export default Service.extend(Ember.Evented, {
     if (this.get('isMobileDevice')) {
       let blockCheck = Ember.run.later(() => {
         this.debug(`Looks like the mobile browser blocked an autoplay trying to play sound with url: ${sound.get('url')}`);
-      }, 1000);
+      }, 2000);
       sound.one('audio-played', () => Ember.run.cancel(blockCheck));
     }
     sound.play();